requirements.txt
supabase-py
python-dotenv
nfc